# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "kasrchem System",
        "description": "kasrchem API Documentation",
        "contact": {
            "email": "admin@kasrchem.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://backend.kasrchem.murabba.dev/api/v2",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get list of Products",
                "operationId": "Get list of Products",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "header",
                        "description": "App Locale",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "ar",
                            "enum": [
                                "ar",
                                "en"
                            ]
                        }
                    },
                    {
                        "name": "section_id",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Show product Data",
                "operationId": "show_products",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "product ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "locale",
                        "in": "header",
                        "description": "App Locale",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "ar",
                            "enum": [
                                "ar",
                                "en"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
...
        "/search": {
            "get": {
                "tags": [
                    "Search"
                ],
                "summary": "Get Search Data",
                "operationId": "get_search_data",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search Word",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "locale",
                        "in": "header",
                        "description": "App Locale",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "ar",
                            "enum": [
                                "ar",
                                "en"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
...
        "/releases": {
            "get": {
                "tags": [
                    "Multimedia"
                ],
                "summary": "Get list of Releases",
                "operationId": "Get list of Releases",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "header",
                        "description": "App Locale",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "ar",
                            "enum": [
                                "ar",
                                "en"
                            ]
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/release_section": {
            "get": {
                "tags": [
                    "Multimedia"
                ],
                "summary": "Get list of sections with release",
                "description": "Returns list of sections with release",
                "operationId": "getReleasList",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "header",
                        "description": "App Locale",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "ar",
                            "enum": [
                                "ar",
                                "en"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "description": "Pass Bearer Token",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}